+++ projects/core/src/server-action.ts
@@ -87,7 +87,7 @@ export class ForbiddenError extends Error {
     constructor() {
         super("Forbidden");
     }
-    isForbiddenError:true = true;
+    isForbiddenError: true = true;
 }
 
 
@@ -125,6 +125,11 @@ export class myServerAction extends Action<inArgs, result>
 export interface BackendMethodOptions<type> {
     /**Determines when this `BackendMethod` can execute, see: [Allowed](https://remult.dev/docs/allowed.html)  */
     allowed: AllowedForInstance<type>;
+    /** Used to determine the route for the BackendMethod.
+     * @example
+     * {allowed:true, apiPrefix:'someFolder/'}
+     */
+    apiPrefix?: string;
     /** EXPERIMENTAL: Determines if this method should be queued for later execution */
     queue?: boolean;
     /** EXPERIMENTAL: Determines if the user should be blocked while this `BackendMethod` is running*/
@@ -135,7 +140,7 @@ export interface BackendMethodOptions<type> {
 export const actionInfo = {
     allActions: [] as any[],
     runningOnServer: false,
-    runActionWithoutBlockingUI:<T> (what: () => Promise<T>): Promise<T> => { return what() },
+    runActionWithoutBlockingUI: <T>(what: () => Promise<T>): Promise<T> => { return what() },
     startBusyWithProgress: () => ({
         progress: (percent: number) => { },
         close: () => { }
@@ -194,7 +199,7 @@ export function BackendMethod<type = any>(options: BackendMethodOptions<type>) {
                 types = options.paramTypes;
             // if types are undefined - you've forgot to set: "emitDecoratorMetadata":true
 
-            let serverAction = new myServerAction(key, types, options, args => originalMethod.apply(undefined, args));
+            let serverAction = new myServerAction((options?.apiPrefix ? (options.apiPrefix + '/') : '') + key, types, options, args => originalMethod.apply(undefined, args));
             serverAction.doWork = async (args, self, url, http) => {
                 args = prepareArgsToSend(types, args);
                 if (options.blockUser === false) {
@@ -243,7 +248,7 @@ export function BackendMethod<type = any>(options: BackendMethodOptions<type>) {
                     }
 
 
-                    reg(controllerOptions.key + '/' + key, options ? options.queue : false, options.allowed, async (d: serverMethodInArgs, req, res) => {
+                    reg(controllerOptions.key + '/' + (options?.apiPrefix ? (options.apiPrefix + '/') : '') + key, options ? options.queue : false, options.allowed, async (d: serverMethodInArgs, req, res) => {
 
                         d.args = d.args.map(x => isCustomUndefined(x) ? undefined : x);
                         let allowed = options.allowed;
@@ -342,7 +347,7 @@ export function BackendMethod<type = any>(options: BackendMethodOptions<type>) {
 
                         let r = await (new class extends Action<serverMethodInArgs, serverMethodOutArgs>{
                             protected execute: (info: serverMethodInArgs, req: Remult, res: DataApiResponse) => Promise<serverMethodOutArgs>;
-                        }(classOptions.key + "/" + key, options ? options.queue : false, options.allowed).run({
+                        }(classOptions.key + "/" + (options?.apiPrefix ? (options.apiPrefix + '/') : '') + key, options ? options.queue : false, options.allowed).run({
                             args,
                             rowInfo: {
                                 data: await defs.toApiJson(),
@@ -366,7 +371,7 @@ export function BackendMethod<type = any>(options: BackendMethodOptions<type>) {
                         await defs.__validateEntity();
                         let r = await (new class extends Action<serverMethodInArgs, serverMethodOutArgs>{
                             protected execute: (info: serverMethodInArgs, req: Remult, res: DataApiResponse) => Promise<serverMethodOutArgs>;
-                        }(x.classes.get(self.constructor).key + "/" + key, options ? options.queue : false, options.allowed).run({
+                        }(x.classes.get(self.constructor).key + "/" + (options?.apiPrefix ? (options.apiPrefix + '/') : '') + key, options ? options.queue : false, options.allowed).run({
                             args,
                             fields: await defs.toApiJson()
                         }, baseUrl, http));