+++ projects/core/src/tests/test-server-controller.spec.ts
@@ -429,6 +429,20 @@ describe("test Server Controller basics", () => {
 
 
     });
+    it("test static backend method ", async () => {
+        const myClass = class {
+            static async doSomething2222() {
+                return isBackend().toString();
+            }
+        }
+        describeClass(myClass,
+            undefined,
+            undefined,
+            {
+                doSomething2222: BackendMethod({ allowed: true ,apiPrefix:'myPrefix'}),
+            })
+        expect(await myClass.doSomething2222()).toBe("true");
+    });
     it("test backend method with adhoc entity", async () => {
         const myClass = class {
             id = 0;
@@ -464,6 +478,7 @@ describe("test Server Controller basics", () => {
         })
         expect(await myClass.adHockDoSomething()).toBe(true);
     });
+
     it("test remult proxy for repository", async () => {
         let remult1 = new Remult(new InMemoryDataProvider());
         await remult1.repo(testEntity).insert({ name: "a" });
@@ -474,13 +489,15 @@ describe("test Server Controller basics", () => {
         expect(await repo.count()).toBe(1);
         remultProxy.remultFactory = () => remult2;
         expect(await repo.count()).toBe(0);
-        let z = await repo.save({name:"b"});
-        expect (z.name).toBe("b");
+        let z = await repo.save({ name: "b" });
+        expect(z.name).toBe("b");
         expect(await repo.count()).toBe(1);
         remultProxy.remultFactory = () => remult1;
-        expect(await repo.find({where:{
-            name:"b"
-        }})).toEqual([])
+        expect(await repo.find({
+            where: {
+                name: "b"
+            }
+        })).toEqual([])
 
     });