+++ projects/test-angular/src/app/server/server.ts
@@ -2,7 +2,7 @@ import { CustomModuleLoader } from './CustomModuleLoader';
 let moduleLoader = new CustomModuleLoader('/dist/test-angular');
 import express, { application } from "express";
 import { remultExpress } from '../../../../core/remult-express'
-import { Task } from "../products-test/products.component";
+import { Task, TasksController, TasksControllerDecorated } from "../products-test/products.component";
 import { JsonDataProvider, Remult, remult } from '../../../../core';
 import { JsonFileDataProvider } from '../../../../core/server';
 import { JobsInQueueEntity } from '../../../../core/server/expressBridge';
@@ -17,12 +17,16 @@ import { createSchema, createYoga } from 'graphql-yoga'
 
 
 
+
+
+
 var r = new Remult();
 r.dataProvider = new JsonFileDataProvider('./db');
 
 const app = express()
 export const api = remultExpress({
     entities: [Task],
+    controllers: [TasksController, TasksControllerDecorated],
     queueStorage: new EntityQueueStorage(r.repo(JobsInQueueEntity))
 })
 app.use(api)
@@ -31,7 +35,7 @@ const openApiDocument = api.openApiDoc({ title: 'remult-react-todo' })
 app.get('/api/openApi.json', (req, res) => res.json(openApiDocument));
 app.use('/api/docs', swaggerUi.serve, swaggerUi.setup(openApiDocument));
 const { schema, rootValue } = remultGraphql(api);
-debugger
+
 app.use('/api/graphql', graphqlHTTP({
     schema: buildSchema(schema),
     rootValue,
@@ -44,7 +48,7 @@ const yoga = createYoga({
         typeDefs: schema,
         resolvers: {
             Query:
-            rootValue
+                rootValue
             // {
             //     tasks: () => [{id:1,title:'noam',completed:false}]
             // }