+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.20.2] - in progress
 - createdAt & updatedAt are by default `allowApiUpdate` false
+- Added `apiPrefix` to `BackendMethod` `options` to allow more control over backend method routes. #189
 ## [0.20.1] - 2023-05-17
 - Improved support for compound id entity
 - Minor bug fixes